LPD-76626 Move existing test to the proper class
