--- modules/apps/site/site-cms-site-initializer-test/src/testIntegration/java/com/liferay/site/cms/site/initializer/internal/model/listener/test/DepotEntryModelListenerTest.java
@@ -1,103 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.site.cms.site.initializer.internal.model.listener.test;
-
-import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.depot.constants.DepotConstants;
-import com.liferay.depot.model.DepotEntry;
-import com.liferay.depot.service.DepotEntryLocalService;
-import com.liferay.document.library.kernel.service.DLFolderLocalService;
-import com.liferay.object.model.ObjectDefinition;
-import com.liferay.object.service.ObjectDefinitionLocalService;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.Repository;
-import com.liferay.portal.kernel.portletfilerepository.PortletFileRepository;
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.service.RepositoryLocalService;
-import com.liferay.portal.kernel.test.rule.AggregateTestRule;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
-import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
-import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.util.TempFileEntryUtil;
-import com.liferay.portal.test.rule.FeatureFlag;
-import com.liferay.portal.test.rule.Inject;
-import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
-import com.liferay.site.cms.site.initializer.test.util.CMSTestUtil;
-
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-/**
- * @author Fábio Alves
- */
-@FeatureFlag("LPD-17564")
-@RunWith(Arquillian.class)
-public class DepotEntryModelListenerTest {
-
-	@ClassRule
-	@Rule
-	public static final AggregateTestRule aggregateTestRule =
-		new AggregateTestRule(
-			new LiferayIntegrationTestRule(),
-			PermissionCheckerMethodTestRule.INSTANCE);
-
-	@Before
-	public void setUp() throws Exception {
-		CMSTestUtil.getOrAddGroup(DepotEntryModelListenerTest.class);
-	}
-
-	@Test
-	public void testOnAfterCreate() throws Exception {
-		DepotEntry depotEntry = _depotEntryLocalService.addDepotEntry(
-			RandomTestUtil.randomLocaleStringMap(),
-			RandomTestUtil.randomLocaleStringMap(), DepotConstants.TYPE_SPACE,
-			ServiceContextTestUtil.getServiceContext());
-
-		Group group = _groupLocalService.fetchGroup(depotEntry.getGroupId());
-
-		ObjectDefinition objectDefinition =
-			_objectDefinitionLocalService.
-				getObjectDefinitionByExternalReferenceCode(
-					"L_CMS_BASIC_DOCUMENT", group.getCompanyId());
-
-		Repository repository = _portletFileRepository.getPortletRepository(
-			group.getGroupId(), objectDefinition.getPortletId());
-
-		Assert.assertNotNull(
-			_dlFolderLocalService.fetchFolder(
-				group.getGroupId(), repository.getDlFolderId(),
-				String.valueOf(TestPropsValues.getUserId())));
-
-		Assert.assertNotNull(
-			_repositoryLocalService.fetchRepository(
-				group.getGroupId(), TempFileEntryUtil.class.getName(),
-				TempFileEntryUtil.class.getName()));
-	}
-
-	@Inject
-	private DepotEntryLocalService _depotEntryLocalService;
-
-	@Inject
-	private DLFolderLocalService _dlFolderLocalService;
-
-	@Inject
-	private GroupLocalService _groupLocalService;
-
-	@Inject
-	private ObjectDefinitionLocalService _objectDefinitionLocalService;
-
-	@Inject
-	private PortletFileRepository _portletFileRepository;
-
-	@Inject
-	private RepositoryLocalService _repositoryLocalService;
-
-}
